moved SingletonSetupTestCase to test_util and added a test for singleton setup itself
